server/requirements.txt
# Core Backend Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1

# Database
SQLAlchemy==2.0.23

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
scipy==1.11.4

# Model Persistence
joblib==1.3.2

# Data Validation
marshmallow==3.20.1

# Reporting
matplotlib==3.8.2
seaborn==0.13.0
reportlab==4.0.7
WeasyPrint==60.1

# Statistics
statsmodels==0.14.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Dysgraphia Image Pipeline
torch>=2.2.0
torchvision>=0.17.0
timm>=1.0.8
pillow>=10.0.0
pyyaml>=6.0.1
